Splitting out implementation into easily browsable units
 - This should help keep the code maintainable as more data objects are added.
