+++ data/boltstorage.go
@@ -2,14 +2,9 @@ package data
 
 import (
 	"encoding/json"
-	"errors"
-	"fmt"
 	"github.com/timshannon/bolthold"
-	"go.etcd.io/bbolt"
 	"os"
 	"path/filepath"
-	"sort"
-	"time"
 )
 
 const (
@@ -44,253 +39,4 @@ func newBoltStorageAt(file string) (*boltStorage, error) {
 		return nil, err
 	}
 	return &boltStorage{store: store}, nil
-}
-
-var _ sort.Interface = (*taskSorter)(nil)
-
-type taskSorter struct {
-	tasks []Task
-}
-
-func newTaskSorter(tasks []Task) *taskSorter {
-	return &taskSorter{tasks: tasks}
-}
-
-func (t *taskSorter) Len() int {
-	return len(t.tasks)
-}
-
-func (t *taskSorter) Less(i, j int) bool {
-	a, b := t.tasks[i], t.tasks[j]
-
-	switch {
-	case !a.Done && b.Done:
-		return true
-	case a.Done && !b.Done:
-		return false
-	case a.Favorite && !b.Favorite:
-		return true
-	case !a.Favorite && b.Favorite:
-		return false
-	case a.Priority > b.Priority:
-		return true
-	case a.Priority < b.Priority:
-		return false
-	case a.ID < b.ID:
-		return true
-	default:
-		return false
-	}
-}
-
-func (t *taskSorter) Swap(i, j int) {
-	t.tasks[i], t.tasks[j] = t.tasks[j], t.tasks[i]
-}
-
-func (b *boltStorage) Get(filters ...TaskFilter) ([]Task, error) {
-	query := bolthold.Where("SoftDeleted").Eq(false)
-	return b.getTasks(query, filters)
-}
-
-func (b *boltStorage) GetHistoric(filters ...TaskFilter) ([]Task, error) {
-	query := new(bolthold.Query)
-	return b.getTasks(query, filters)
-}
-
-func (b *boltStorage) getTasks(query *bolthold.Query, filters []TaskFilter) ([]Task, error) {
-	for _, f := range filters {
-		f(query)
-	}
-
-	var found []Task
-	err := b.store.Bolt().View(func(tx *bbolt.Tx) error {
-		if err := b.store.TxFind(tx, &found, query); err != nil {
-			return err
-		}
-		return nil
-	})
-	if err != nil {
-		return nil, err
-	}
-	sort.Sort(newTaskSorter(found))
-	return found, nil
-}
-
-func (b *boltStorage) Count() (int, error) {
-	var count int
-	err := b.store.Bolt().View(func(tx *bbolt.Tx) error {
-		var err error
-		count, err = b.store.TxCount(tx, new(Task), bolthold.Where("SoftDeleted").Eq(false))
-		if err != nil {
-			return err
-		}
-		return nil
-	})
-	if err != nil {
-		return 0, err
-	}
-	return count, nil
-}
-
-func (b *boltStorage) Create(task Task) (Task, error) {
-	err := b.store.Bolt().Update(func(tx *bbolt.Tx) error {
-		if err := b.store.TxInsert(tx, bolthold.NextSequence(), &task); err != nil {
-			return err
-		}
-		return nil
-	})
-	if err != nil {
-		return Task{}, err
-	}
-	return task, nil
-}
-
-func (b *boltStorage) Update(id uint64, task Task) (Task, error) {
-	err := b.store.Bolt().Update(func(tx *bbolt.Tx) error {
-		return b.store.TxUpdate(tx, id, &task)
-	})
-	if err != nil {
-		if errors.Is(err, bolthold.ErrNotFound) {
-			return Task{}, fmt.Errorf("%w: %v", ErrIDNotFound, err)
-		}
-		return Task{}, err
-	}
-	return task, nil
-}
-
-func (b *boltStorage) Delete(id uint64) error {
-	err := b.store.Bolt().Update(func(tx *bbolt.Tx) error {
-		task := new(Task)
-		if err := b.store.TxFindOne(tx, task, bolthold.Where(bolthold.Key).Eq(id)); err != nil {
-			return nil
-		}
-		task.SoftDeleted = true
-		return b.store.TxUpdate(tx, id, task)
-	})
-	if err != nil {
-		if errors.Is(err, bolthold.ErrNotFound) {
-			return nil
-		}
-		return err
-	}
-	return nil
-}
-
-func (b *boltStorage) Export(dir string) (outName string, err error) {
-	out, err := os.CreateTemp(dir, exportPattern)
-	if err != nil {
-		return "", err
-	}
-	defer func() {
-		_ = out.Close()
-	}()
-
-	outName = out.Name()
-
-	writer := json.NewEncoder(out)
-
-	tasks, err := b.GetHistoric()
-	if err != nil {
-		return "", err
-	}
-	entries, err := b.GetTimeEntries()
-	if err != nil {
-		return "", err
-	}
-	err = writer.Encode(exportModel{Tasks: tasks, TimeEntries: entries})
-	if err != nil {
-		return
-	}
-
-	return outName, nil
-}
-
-func (b *boltStorage) StartTimeEntry(taskID uint64) (TimeEntry, error) {
-	var entry TimeEntry
-	err := b.store.Bolt().Update(func(tx *bbolt.Tx) error {
-		now := time.Now()
-		var err error
-		entry, err = b.txStartTimeEntry(tx, taskID, now)
-		if err != nil {
-			return err
-		}
-		return nil
-	})
-	if err != nil {
-		return TimeEntry{}, err
-	}
-	return entry, nil
-}
-
-func (b *boltStorage) txStartTimeEntry(tx *bbolt.Tx, taskID uint64, now time.Time) (TimeEntry, error) {
-	task := new(Task)
-	if err := b.store.TxFindOne(tx, task, bolthold.Where(bolthold.Key).Eq(taskID)); err != nil {
-		return TimeEntry{}, err
-	}
-	entry := TimeEntry{
-		TaskID: taskID,
-		Start:  now,
-	}
-	if err := b.store.TxInsert(tx, bolthold.NextSequence(), &entry); err != nil {
-		return TimeEntry{}, err
-	}
-	return entry, nil
-}
-
-func (b *boltStorage) StopTimeEntry(entryID uint64) error {
-	now := time.Now()
-
-	err := b.store.Bolt().Update(func(tx *bbolt.Tx) error {
-		return b.txStopTimeEntry(tx, entryID, now)
-	})
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func (b *boltStorage) txStopTimeEntry(tx *bbolt.Tx, entryID uint64, now time.Time) error {
-	found := new(TimeEntry)
-	if err := b.store.TxFindOne(tx, found, bolthold.Where(bolthold.Key).Eq(entryID)); err != nil {
-		return err
-	}
-	found.End = &now
-	return b.store.TxUpdate(tx, entryID, *found)
-}
-
-func (b *boltStorage) StartAfterStop(startTaskID uint64, stopEntryID uint64) (TimeEntry, error) {
-	now := time.Now()
-	var entry TimeEntry
-
-	err := b.store.Bolt().Update(func(tx *bbolt.Tx) error {
-		if err := b.txStopTimeEntry(tx, stopEntryID, now); err != nil {
-			return err
-		}
-		_entry, err := b.txStartTimeEntry(tx, startTaskID, now)
-		if err != nil {
-			return err
-		}
-		entry = _entry
-		return nil
-	})
-	return entry, err
-}
-
-func (b *boltStorage) GetTimeEntries(filters ...TimeEntryFilter) ([]TimeEntry, error) {
-	var entries []TimeEntry
-	query := new(bolthold.Query)
-
-	for _, filter := range filters {
-		filter(query)
-	}
-
-	query.SortBy("Start")
-
-	err := b.store.Bolt().Update(func(tx *bbolt.Tx) error {
-		return b.store.TxFind(tx, &entries, query)
-	})
-	if err != nil {
-		return nil, err
-	}
-	return entries, err
 }
