+++ data/data.go
@@ -1,2 +1,3 @@
 // Package data provides base types and structure for data interactions.
+// This currently includes Task and TimeEntry data.
 package data
